2025/03/14 change to monolith for development

+++ src/main/java/com/jack/aquark/dto/RawDataItemDto.java
@@ -3,6 +3,7 @@ package com.jack.aquark.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -15,7 +16,7 @@ public class RawDataItemDto {
   private String csq;
 
   private Sensor sensor;
-  private Double rainD;
+  private BigDecimal rainD;
 
   @Data
   public static class Sensor {
@@ -34,28 +35,28 @@ public class RawDataItemDto {
 
   @Data
   public static class Volt {
-    private Double v1;
-    private Double v2;
-    private Double v3;
-    private Double v4;
-    private Double v5;
-    private Double v6;
-    private Double v7;
+    private BigDecimal v1;
+    private BigDecimal v2;
+    private BigDecimal v3;
+    private BigDecimal v4;
+    private BigDecimal v5;
+    private BigDecimal v6;
+    private BigDecimal v7;
   }
 
   @Data
   public static class StickTxRh {
-    private Double rh;
-    private Double tx;
+    private BigDecimal rh;
+    private BigDecimal tx;
   }
 
   @Data
   public static class UltrasonicLevel {
-    private Double echo;
+    private BigDecimal echo;
   }
 
   @Data
   public static class WaterSpeedAquark {
-    private Double speed;
+    private BigDecimal speed;
   }
 }
